URLEncode query arguments and add toQuery
Use URLEncoder to encode filter arguments because application tags with spaces weren't being queried properly.

Change toString to toQuery mostly because URLEncoder throws UnsupportedEncodingException and we can't throw exceptions
from toString because it overrides Object.toString. This change also makes sense because we're not really generating
a human-readable representation of the form.