🚧 JAVA-3298 Resource Types Delegate To Value Objects
GSON cannot deserialize a resource type, because resource type constructors require a GSON instance, and GSON registers type adapters at creation thus creating a circular dependency. We could break this by setting the GSON instance of the type adapters after constructing the GSON instance, but this is kind of gross. Instead, we delegate JSON deserialization to a package-private value type, and resource types delegate accessors to the value type.